feat: add HBM3 PuD primitive substrate

+++ src/ramulator/controller/controller_base.h
@@ -147,7 +147,7 @@ class ControllerBase : public IController, public Implementation {
 
   // Cached spec lookups
   int m_bank_level = -1;
-  int m_tCK_ps = -1;
+  double m_tCK_ps = -1;
 
   // Per flat-bank count of requests in m_active_buffer (typically 0 or 1).
   // Maintained by promote_to_active / retire_request.
